==> code/02 feature extraction - lemmas , POS, dependency - using spaCy.ipynb <==
import pandas as pd
import spacy
from tqdm import tqdm_notebook as tqdm
import csv

csv.field_size_limit(100000000)

# Read csvs from before.

test_df = pd.read_csv('data/test_df.csv', encoding='utf8', engine='python') 

# Training dataset is too large - working in chunks.
train_df = pd.read_csv('data/train_df.csv', encoding='utf8', engine='python') 

byarticle_df = pd.read_csv('byarticle_df.csv', encoding='utf8', engine='python')

# Lemmatize using spaCy

All three given datasets are lemmatized, tagged and saved into new pandas datasets.

# Handle NaN

def check_str(text):
    if type(text) != str:
        print(text)
        return ''
    else:
        return text

#test_df['text'] = test_df['text'].apply(check_str)

nlp = spacy.load('en', disable=['ner','textcat']) # skipped NER, could come back

def parse(df, col):
    lemma = []
    pos = []
    tag = []
    dep = []
    for doc in tqdm(nlp.pipe(df[col], batch_size=50, n_threads=3)):
        lemma.append(tuple([n.lemma_ for n in doc]))
        pos.append(tuple([n.pos_ for n in doc]))
        tag.append(tuple([n.tag_ for n in doc]))
        dep.append(tuple([n.dep_ for n in doc]))
        
    df[col + '_lemmas'] = lemma
    df[col + '_pos'] = pos
    df[col + '_tags'] = tag
    df[col + '_deps'] = dep
    
    return df

test_df = test_df.drop(columns=['text'])
train_df = train_df.drop(columns=['text'])

#test_df = test_df.dropna()
test_df = parse(test_df, 'title')
test_df.head()

test_df.to_csv('data/test_df_titles_spacy.csv', index=False)

train_df = train_df.dropna()

train_df = parse(train_df, 'title')
train_df.head()

train_df.to_csv('data/train_df_titles_spacy.csv', index=False)

# Training dataset is too large. Do in chunks.
i = 1
for chunk in pd.read_csv('data/train_df.csv', chunksize=100000, encoding='utf8', engine='python'):
    print('Processing chunk ' + str(i))
    chunk['text'] = chunk['text'].apply(check_str)
    chunk = parse(chunk)
    chunk.to_csv('train_df_spacy' + str(i) + '.csv', index=False)
    i += 1

train_df.head()

train_df.to_csv('data/train_df_spacy.csv', index=False)

byarticle_df = parse(byarticle_df)

byarticle_df.head()

byarticle_df.to_csv('byarticle_df_spacy.csv', index=False)
