==> code/03 bag-of-words unigram.ipynb <==
import pandas as pd
import numpy as np
import sklearn as sk
from tqdm import tqdm
import ast
import csv
import glob
import pickle

csv.field_size_limit(100000000)

# Read in csv datasets

files = glob.glob('data/train_df_spacy*.csv')

# attempting on a .1 random sampled set

test_df = pd.read_csv('data/test_df_spacy.csv').sample(frac=.1, random_state=123)

frames = []
for file in files:
    frames.append(pd.read_csv(file).sample(frac=.1, random_state=123))
train_df = pd.concat(frames)

test_df.head()

train_df.head()

# setup for initial model experimentation

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.svm import LinearSVC

def dummy(doc):
    # use as param in vectorizers
    return ast.literal_eval(doc)

count_uni = CountVectorizer(stop_words = 'english',
                    analyzer='word',
                    tokenizer=dummy,
                    preprocessor=None,
                    max_features = 5000)
tfidf_uni = TfidfVectorizer(stop_words = 'english',
                    analyzer='word',
                    tokenizer=dummy,
                    preprocessor=None,
                    max_features = 5000)
count_bi = CountVectorizer(stop_words = 'english',
                    analyzer='word',
                    tokenizer=dummy,
                    preprocessor=None,
                    ngram_range=(1,2),
                    max_features = 10000)
tfidf_bi = TfidfVectorizer(stop_words = 'english',
                    analyzer='word',
                    tokenizer=dummy,
                    preprocessor=None,
                    ngram_range=(1,2),
                    max_features = 10000)

dt = DecisionTreeClassifier()
lr = LogisticRegression(verbose=True)
svm = LinearSVC(verbose=True)
mlp = MLPClassifier(verbose=True)

def model(model, vectorizer, x_train, y_train, x_test, y_test):
    # train a model (can be various kinds) & evaluate
    
    x_train_use = vectorizer.fit_transform(x_train) # if using dask, add x_train.compute()
    x_test_use = vectorizer.transform(x_test)
    
    model = model.fit(x_train_use, y_train)
    y_pred = model.predict(x_test_use)
    
    acc = sk.metrics.accuracy_score(y_test, y_pred)
    print("accuracy: {0:.2f}%".format(acc*100))
    
    mapping = list(zip(vectorizer.get_feature_names(), model.coef_[0]))
    mapping.sort(key=lambda x: abs(x[1]))
    mapping.reverse()
    
    return mapping

y_train = train_df['label']
y_test = test_df['label']

x_train = train_df['title_lemmas']
x_test = test_df['title_lemmas']

# logistic regression + count vectors: pre-vocab cut

lr_count_uni = model(lr, count_uni, x_train, y_train, x_test, y_test)

lr_count_uni

# test on own train set just to see
test_on_train = model(lr, count_uni, x_train, y_train, x_train, y_train)

# test on other sample of training set
test_train = pd.read_csv('data/train_df_spacy1.csv').sample(frac=.1, random_state=1)
y_test_train = test_train['label']
x_test_train = test_train['text_lemmas']
test_on_other_train = model(lr, count_uni, x_train, y_train, x_test_train, y_test_train)

# pre-vocab-cut with TFIDF instead of count

lr_tfidf_uni = model(lr, tfidf_uni, x_train, y_train, x_test, y_test)

lr_tfidf_uni

# remove tokens overrepresented in the dataset from vocabulary (they are likely publisher-related)

See 4.1 in the paper.

def examine_tokens(df, token):
    num = 0
    for index, row in df.iterrows():
        if token in row['text_lemmas']:
            num += 1
            if num <= 20:
                print(row['text'])
    return num

def token_disparity(x_train, x_test, tokens):
    for token in tokens:
        token = str(token[0])
        train_count = 0
        for train_item in x_train:
            train_count += train_item.count(token)
        test_count = 0
        for test_item in x_test:
            test_count += test_item.count(token)
        train_norm = train_count/len(x_train)
        test_norm = test_count/len(x_test)
        if test_norm == 0:
            val = 'does not appear in test'
            print(token)
        else:
            val = train_norm/test_norm
            if val > 5:
                print(token)

token_disparity(x_train, x_test, lr_count_uni)

token_disparity(x_train, x_test, mapping)

# upgraded vectorization - own counter

from collections import Counter
import re

def dummy_two(doc):
    # use as param in vectorizers
    return doc

def countvocab(text_lemmas):
    
    vocab = Counter()
    word = re.compile('[a-z]*\Z')
    
    except_file = open('token_exceptions.txt')
    exceptions = except_file.read().split('\n')
    except_file.close()
    stopwords_file = open('sklearn_stopwords.txt')
    stopwords = stopwords_file.read().split('\n')
    stopwords_file.close()
    
    for text in text_lemmas:
        for token in ast.literal_eval(text):
            if word.match(token) == None:
                continue
            if token in exceptions:
                continue
            if token in stopwords:
                continue
            vocab[token] += 1
    
    return [couple[0] for couple in vocab.most_common(4000)]

vocab = countvocab(x_train)

vocab_file = open('vocab.txt', 'w')
for word in vocab:
    vocab_file.write(word+'\n')
vocab_file.close()

count_uni = CountVectorizer(vocabulary = vocab,
                    analyzer='word',
                    tokenizer=dummy_two,
                    preprocessor=None,
                    lowercase = False)

train_fvs = count_uni.transform(x_train)

test_fvs = count_uni.transform(x_test)

def count_fvs(text):
    
    text = ast.literal_eval(text)
    
    vector = [0]*len(vocab)
    
    for idx in range(len(vocab)):
        vector[idx] = text.count(vocab[idx])
    
    return vector

train_fvs = x_train.apply(count_fvs)
test_fvs = x_test.apply(count_fvs)

# Unigram logistic regression with updated vocabulary

model = lr.fit(train_fvs, y_train)
y_pred = model.predict(test_fvs)
    
acc = sk.metrics.accuracy_score(y_test, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))
    
mapping = list(zip(vocab, model.coef_[0]))
mapping.sort(key=lambda x: abs(x[1]))
mapping.reverse()

mapping

pickle.dump(model, open('hyperpartisan-model.sav', 'wb'))

# Evaluation

# Read datasets, read hyperpartisan model, give percentages.
