==> code/04 hand-crafted feature extraction and model training.ipynb <==
import pandas as pd
import nltk
import spacy
from glob import glob
import sklearn as sk
import ast
import csv
import re
from collections import Counter
from collections import defaultdict

csv.field_size_limit(100000000)

# Feature Extraction

These are the functions used to extract the linguistic features from the spaCy tags previously created.

There are also functions used to extract counts of words from specific lexica that have been referenced.

# POS tag counts

def count_adj(pos):
    # takes unevaluated list of POS tags for a single text
    pos = ast.literal_eval(pos)
    if len(pos) == 0:
        return 0.0
    return pos.count('ADJ')/len(pos)

def count_adv(pos):
    # takes unevaluated list of POS tags for a single text
    pos = ast.literal_eval(pos)
    if len(pos) == 0:
        return 0.0
    return pos.count('ADV')/len(pos)

def count_pron(pos):
    # takes unevaluated list of POS tags for a single text
    pos = ast.literal_eval(pos)
    if len(pos) == 0:
        return 0.0
    return pos.count('PRON')/len(pos)

def count_noun(pos):
    # takes unevaluated list of POS tags for a single text
    pos = ast.literal_eval(pos)
    if len(pos) == 0:
        return 0.0
    return pos.count('NOUN')/len(pos)

def count_propn(pos):
    # takes unevaluated list of POS tags for a single text
    pos = ast.literal_eval(pos)
    if len(pos) == 0:
        return 0.0
    return pos.count('PROPN')/len(pos)

# MPQA lexicons: intensifiers, subjectivity clues (strong, weak)

with open('lexica/intensifiers.tff') as f:
    intensifiers = [line[29:-15].split(' ')[0] for line in f.readlines()]
    
with open('lexica/subjclueslen1-HLTEMNLP05.tff') as f:
    clues = [(line[5:].split(' ')[0], line[5:].split(' ')[2][6:]) for line in f.readlines()]
    strongsubj = []
    weaksubj = []
    for strength, word in clues:
        if strength == 'weaksubj':
            weaksubj.append(word)
        else:
            strongsubj.append(word)
            
def count_strongsubj(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in strongsubj:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

def count_weaksubj(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in weaksubj:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

# count passives, modality

def count_aux(dep):
    dep = ast.literal_eval(dep)
    if len(dep) == 0:
        return 0.0
    return (dep.count('aux') + dep.count('auxpass')) / len(dep)

def count_pass(dep):
    dep = ast.literal_eval(dep)
    if len(dep) == 0:
        return 0.0
    return (dep.count('nsubjpass') + dep.count('csubjpass')) / len(dep)

# readability features

def avg_sent_length(text):
    if type(text) != str:
        return 0.0
    words = nltk.word_tokenize(text)
    sents = nltk.sent_tokenize(text)
    if len(sents) == 0:
        return 0.0
    return len(words)/len(sents)

def avg_word_length(text):
    if type(text) != str:
        return 0.0
    words = nltk.word_tokenize(text)
    if len(words) == 0:
        return 0.0
    return len(text)/len(words)

def exclamation_marks(text):
    if type(text) != str:
        return 0.0
    if len(text) == 0:
        return 0.0
    return text.count('!')/len(text)

def question_marks(text):
    if type(text) != str:
        return 0.0
    if len(text) == 0:
        return 0.0
    return text.count('?')/len(text)

def multiple_punct(text):
    if type(text) != str:
        return 0.0
    if len(text) == 0:
        return 0.0
    count = 0
    count += text.count('!!')
    count += text.count('??')
    count += text.count('?!')
    count += text.count('!?')
    return count/len(text)

# EmoLex

emotions = defaultdict(lambda: [])

with open('lexica/NRC-Sentiment-Emotion-Lexicons/NRC-Affect-Intensity-Lexicon/NRC-AffectIntensity-Lexicon.txt') as f:
    pairs = [(line[:-1].split('\t')[0], line[:-1].split('\t')[2]) for line in f.readlines()[1:]]
    for word, emotion in pairs:
        emotions[emotion].append(word)

def count_anger(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in emotions['anger']:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

def count_fear(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in emotions['fear']:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

def count_sadness(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in emotions['sadness']:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

def count_joy(lemmas):
    lemmas = ast.literal_eval(lemmas)
    if len(lemmas) == 0:
        return 0.0
    clues = 0
    for word in emotions['joy']:
        if word in lemmas:
            clues += 1
    return clues/len(set(lemmas))

# Hapax Legomena

def countvocab(df):
    
    vocab = Counter()
    word = re.compile('[a-z]*\Z')
    
    #except_file = open('token_exceptions.txt')
    #exceptions = except_file.read().split('\n')
    #except_file.close()
    #stopwords_file = open('sklearn_stopwords.txt')
    #stopwords = stopwords_file.read().split('\n')
    #stopwords_file.close()
    
    for text in df['text_lemmas']:
        for token in ast.literal_eval(text):
            if word.match(token) == None:
                continue
            #if token in exceptions:
            #    continue
            #if token in stopwords:
            #    continue
            vocab[token] += 1
            
    hapax_legomena = [word for word in vocab if vocab[word]==1]
            
    true = 0
    true_total = 0
    false = 0
    false_total = 0
    
    for row, item in df.iterrows():
        if item['label'] == True:
            true_total += 1
        else:
            false_total += 1
            
        if len(list(set(hapax_legomena).intersection(ast.literal_eval(item['text_lemmas'])))) != 0:
            if item['label'] == True:
                true += 1
            else:
                false += 1
    
    print('Proportion of biased articles with hapax legomena: ' + str(float(true) / true_total))
    print('Proportion of unbiased articles with hapax legomena: ' + str(float(false) / false_total))
    
    return hapax_legomena

countvocab(train_df)

# Testing

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.svm import LinearSVC
from sklearn.linear_model import RidgeClassifier

feature_cols = ['ADJ_count', 'ADV_count', 'PRON_count', 'NOUN_count', 'PROPN_count', 
                'strongsubj_count', 'weaksubj_count', 'aux_count', 'pass_count',
               'sent_length', 'word_length', 'excl_count', 'quest_count', 'extra_punct_count',
               'anger_count', 'fear_count', 'sadness_count', 'joy_count']

## Small article dataset

df = pd.read_csv('byarticle_df_spacy.csv', encoding='utf8', engine='python')

# balance classes in byarticle dataset (238, 242 -> 480)
biased = df.loc[df.label == True]
unbiased = df.loc[df.label == False]
df = pd.concat([biased, unbiased.sample(n=242).reset_index(drop=True)])

feature_df = pd.DataFrame()
feature_df['label'] = df.label

feature_df['ADJ_count'] = df['text_pos'].apply(count_adj)
feature_df['ADV_count'] = df['text_pos'].apply(count_adv)
feature_df['PRON_count'] = df['text_pos'].apply(count_pron)
feature_df['NOUN_count'] = df['text_pos'].apply(count_noun)
feature_df['PROPN_count'] = df['text_pos'].apply(count_propn)

feature_df['strongsubj_count'] = df['text_lemmas'].apply(count_strongsubj)
feature_df['weaksubj_count'] = df['text_lemmas'].apply(count_weaksubj)

feature_df['aux_count'] = df['text_deps'].apply(count_aux)
feature_df['pass_count'] = df['text_deps'].apply(count_pass)

feature_df['sent_length'] = df['text'].apply(avg_sent_length)
feature_df['word_length'] = df['text'].apply(avg_word_length)
feature_df['excl_count'] = df['text'].apply(exclamation_marks)
feature_df['quest_count'] = df['text'].apply(question_marks)
feature_df['extra_punct_count'] = df['text'].apply(multiple_punct)

feature_df['anger_count'] = df['text_lemmas'].apply(count_anger)
feature_df['fear_count'] = df['text_lemmas'].apply(count_fear)
feature_df['sadness_count'] = df['text_lemmas'].apply(count_sadness)
feature_df['joy_count'] = df['text_lemmas'].apply(count_joy)

feature_df.head()

# you want all rows, and the feature_cols' columns
X = feature_df.loc[:, feature_cols]
X.shape

y = feature_df.label
y.shape

from sklearn.model_selection import train_test_split
X_tr, X_te, y_tr, y_te = train_test_split(X, y, test_size=0.2)

lr_small = RidgeClassifier()

lr_small.fit(X, y)
y_pred = lr_small.predict(X)
acc = sk.metrics.accuracy_score(y, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

y_pred = lr_small.predict(X_train)
acc = sk.metrics.accuracy_score(y_train, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

y_pred = lr_small.predict(X_test)
acc = sk.metrics.accuracy_score(y_test, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

mapping = list(zip(feature_cols, lr_small.coef_[0]))
mapping

## Complete dataset

train_files = glob('data/train_df_spacy*.csv')

test_df = pd.read_csv('data/test_df_spacy.csv').sample(frac=.01, random_state=123)

frames = []
for file in train_files:
    frames.append(pd.read_csv(file).sample(frac=.01, random_state=123))
train_df = pd.concat(frames)

train_feature_df = pd.DataFrame()
train_feature_df['label'] = train_df.label
test_feature_df = pd.DataFrame()
test_feature_df['label'] = test_df.label

train_feature_df['ADJ_count'] = train_df['text_pos'].apply(count_adj)
train_feature_df['ADV_count'] = train_df['text_pos'].apply(count_adv)
train_feature_df['PRON_count'] = train_df['text_pos'].apply(count_pron)
train_feature_df['NOUN_count'] = train_df['text_pos'].apply(count_noun)
train_feature_df['PROPN_count'] = train_df['text_pos'].apply(count_propn)

train_feature_df['strongsubj_count'] = train_df['text_lemmas'].apply(count_strongsubj)
train_feature_df['weaksubj_count'] = train_df['text_lemmas'].apply(count_weaksubj)

train_feature_df['aux_count'] = train_df['text_deps'].apply(count_aux)
train_feature_df['pass_count'] = train_df['text_deps'].apply(count_pass)

train_feature_df['sent_length'] = train_df['text'].apply(avg_sent_length)
train_feature_df['word_length'] = train_df['text'].apply(avg_word_length)
train_feature_df['excl_count'] = train_df['text'].apply(exclamation_marks)
train_feature_df['quest_count'] = train_df['text'].apply(question_marks)
train_feature_df['extra_punct_count'] = train_df['text'].apply(multiple_punct)

train_feature_df['anger_count'] = train_df['text_lemmas'].apply(count_anger)
train_feature_df['fear_count'] = train_df['text_lemmas'].apply(count_fear)
train_feature_df['sadness_count'] = train_df['text_lemmas'].apply(count_sadness)
train_feature_df['joy_count'] = train_df['text_lemmas'].apply(count_joy)

test_feature_df['ADJ_count'] = test_df['text_pos'].apply(count_adj)
test_feature_df['ADV_count'] = test_df['text_pos'].apply(count_adv)
test_feature_df['PRON_count'] = test_df['text_pos'].apply(count_pron)
test_feature_df['NOUN_count'] = test_df['text_pos'].apply(count_noun)
test_feature_df['PROPN_count'] = test_df['text_pos'].apply(count_propn)

test_feature_df['strongsubj_count'] = test_df['text_lemmas'].apply(count_strongsubj)
test_feature_df['weaksubj_count'] = test_df['text_lemmas'].apply(count_weaksubj)

test_feature_df['aux_count'] = test_df['text_deps'].apply(count_aux)
test_feature_df['pass_count'] = test_df['text_deps'].apply(count_pass)

test_feature_df['sent_length'] = test_df['text'].apply(avg_sent_length)
test_feature_df['word_length'] = test_df['text'].apply(avg_word_length)
test_feature_df['excl_count'] = test_df['text'].apply(exclamation_marks)
test_feature_df['quest_count'] = test_df['text'].apply(question_marks)
test_feature_df['extra_punct_count'] = test_df['text'].apply(multiple_punct)

test_feature_df['anger_count'] = test_df['text_lemmas'].apply(count_anger)
test_feature_df['fear_count'] = test_df['text_lemmas'].apply(count_fear)
test_feature_df['sadness_count'] = test_df['text_lemmas'].apply(count_sadness)
test_feature_df['joy_count'] = test_df['text_lemmas'].apply(count_joy)

train_feature_df.head()

X_train = train_feature_df.loc[:, feature_cols]
X_test = test_feature_df.loc[:, feature_cols]
y_train = train_feature_df.label
y_test = test_feature_df.label

lr = RidgeClassifier()

lr.fit(X_train, y_train)
y_pred = lr.predict(X_test)
acc = sk.metrics.accuracy_score(y_test, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

y_pred = lr.predict(X_train)
acc = sk.metrics.accuracy_score(y_train, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

y_pred = lr.predict(X)
acc = sk.metrics.accuracy_score(y, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

mapping = list(zip(feature_cols, lr.coef_[0]))
mapping

from sklearn.model_selection import cross_val_score

X_total = pd.concat([X_train.sample(frac=.25, random_state=123), X_test])
y_total = [0]*len(X_test) + [1]*len(X_test)

scores = cross_val_score(lr, X_total, y_total, cv=15)

scores

sum(scores)/len(scores)

lr.fit(X_test, y_test)
y_pred = lr.predict(X_train)
acc = sk.metrics.accuracy_score(y_train, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))

y_pred = lr.predict(X_test)
acc = sk.metrics.accuracy_score(y_test, y_pred)
print("accuracy: {0:.2f}%".format(acc*100))
