==> code/05 train word embeddings.ipynb <==
import numpy as np
import gensim
from nltk.tokenize import sent_tokenize, word_tokenize
import pandas as pd
import pickle

This notebook is used to train word2vec embeddings with gensim over the given training and validation datasets.

#titles = pd.read_csv('df/train_df.csv')['title']
#texts = pd.read_csv('df/train_df.csv')['text']
titles = pd.read_csv('df/test_df.csv')['title']
texts = pd.read_csv('df/test_df.csv')['text']

sentences = []
for title in titles:
    if type(title) == str:
        sentences.append(word_tokenize(title))

total = len(texts)
for i in range(len(texts)):
    if i % 10000 == 0:
        print(i)
    text = texts[i]
    if type(text) == str:
        text = sent_tokenize(text)
        for sent in text:
            sentences.append(word_tokenize(sent))

with open('sentences-test.pkl', 'wb') as f:
    pickle.dump(sentences, f)
with open('sentences.pkl', 'rb') as f:
    sentences = sentences + pickle.load(f)

embed_dim = 50

model = gensim.models.Word2Vec(sentences, 
                               size = embed_dim, 
                               window = 5, 
                               min_count = 5, 
                               negative = 15,
                               iter = 2, 
                               workers = 12)

model.most_similar(positive=['trump'])

word2index = {token: token_index for token_index, token in enumerate(model.wv.index2word)} 

pickle.dump(word2index, open("embed_total_idx.pkl", "wb"))

embedding_matrix = np.zeros((len(model.wv.vocab), embed_dim))
for i in range(len(model.wv.vocab)):
    embedding_vector = model[model.wv.index2word[i]]
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector

np.save('embed_total_vec', embedding_matrix)

embedding_matrix.shape
