==> code/06 LSTMs.ipynb <==
import tensorflow as tf
import numpy as np
from tensorflow.contrib import rnn
from random import randint
import gensim
import pickle
from sklearn.model_selection import train_test_split
from nltk.tokenize import word_tokenize
from sklearn.preprocessing import OneHotEncoder
import pandas as pd

This notebook shows the development of an LSTM model using TensorFlow. Data is read from preprocessed texts in a CSV, transformed into the appropriate form for the model, then fed into an LSTM.

Note that due to computational constraints, this was entirely done on the University of Potsdam GPU servers.

# Import Data

%cd '/data/hyperpartisan-news-detection'

X_test = pd.read_csv('df/test_df.csv')['text']
y_test = pd.read_csv('df/test_df.csv')['label']
X_train = pd.read_csv('df/train_df.csv')['text']
y_train = pd.read_csv('df/train_df.csv')['label']
X_byarticle = pd.read_csv('df/byarticle_df.csv')['text']
y_byarticle = pd.read_csv('df/byarticle_df.csv')['label']

X_test = pd.read_csv('df/test_agree_df.csv')['text']
y_test = pd.read_csv('df/test_agree_df.csv')['label']
X_train = pd.read_csv('df/train_agree_df.csv')['text']
y_train = pd.read_csv('df/train_agree_df.csv')['label']
X_byarticle = pd.read_csv('df/byarticle_df.csv')['text']
y_byarticle = pd.read_csv('df/byarticle_df.csv')['label']
#X_byarticle = pd.read_csv('df/byarticle_agree_df.csv')['text']
#y_byarticle = pd.read_csv('df/byarticle_agree_df.csv')['label']

# finish data input

X_train_main, X_train_holdout, y_train_main, y_train_holdout = train_test_split(X_train, y_train, test_size=0.2, random_state=100)

enc = OneHotEncoder(handle_unknown='ignore', sparse=False)
y_test = np.array(enc.fit_transform(np.array(y_test).reshape(-1, 1)))
y_train = np.array(enc.fit_transform(np.array(y_train).reshape(-1, 1)))
#y_train_main = np.array(enc.fit_transform(np.array(y_train_main).reshape(-1, 1)))
#y_train_holdout = np.array(enc.fit_transform(np.array(y_train_holdout).reshape(-1, 1)))
y_byarticle = np.array(enc.fit_transform(np.array(y_byarticle).reshape(-1, 1)))
#y_clean = np.array(enc.fit_transform(np.array(y_clean).reshape(-1, 1)))

#X_total = np.array(X_total)
#X_clean = np.array(X_clean)
X_train = np.array(X_train)
X_test = np.array(X_test)
#X_train_main = np.array(X_train_main)
#X_train_holdout = np.array(X_train_holdout)
X_byarticle = np.array(X_byarticle)

X_total = np.concatenate((X_train, X_test, X_byarticle))
y_total = np.concatenate((y_train, y_test, y_byarticle))

## Prepare word embedding matrices (one-time necessity)

words = []
idx = 0
word2idx = {}
vectors = np.zeros((400000,100))

#with open('word-embeddings/glove.6B.50d.txt') as f:
with open('word-embeddings/glove.6B.100d.txt') as f:
    for l in f:
        line = l.split()
        word = line[0]
        words.append(word)
        word2idx[word] = idx
        vect = np.array(line[1:]).astype(np.float)
        vectors[idx] = vect
        idx += 1

vectors.shape

# add zero vector for padding
#vectors = np.append(vectors, np.zeros([1, 50]), axis=0)
vectors = np.append(vectors, np.zeros([1, 100]), axis=0)
vectors.shape

#pickle.dump(word2idx, open("word-embeddings/glove_indices.pkl", "wb"))
pickle.dump(word2idx, open("word-embeddings/glove100_indices.pkl", "wb"))

#np.save('word-embeddings/glove_vectors', vectors)
np.save('word-embeddings/glove100_vectors', vectors)

model = gensim.models.KeyedVectors.load_word2vec_format('word-embeddings/GoogleNews-vectors-negative300.bin', binary=True) 

word2index = {token: token_index for token_index, token in enumerate(model.index2word)} 

pickle.dump(word2index, open("word-embeddings/word2vec_indices.pkl", "wb"))

embedding_matrix = np.zeros((len(model.vocab), 300))
for i in range(len(model.vocab)):
    embedding_vector = model[model.index2word[i]]
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector

np.save('word-embeddings/embedding_matrix', embedding_matrix)

# Model

batch_size = 16
lstm_units = 50
num_classes = 2
max_sent_length = 100
iterations = 16000 #8k for train-main, 12k for total

#embedding_matrix = np.load('word-embeddings/glove100_vectors.npy')
#word2idx = pickle.load(open("word-embeddings/glove100_indices.pkl", "rb" ))
embedding_matrix = np.load('word-embeddings/embed_total_vec.npy')
word2idx = pickle.load(open("word-embeddings/embed_total_idx.pkl", "rb" ))

word_embed_size = embedding_matrix.shape[1]
embed_vocab_size = embedding_matrix.shape[0]

embedding_matrix = np.append(embedding_matrix, np.zeros([1, word_embed_size]), axis=0)
embedding_matrix.shape

def nextBatch(batch_size, data, labels):
    '''
    Return a total of `num` random samples and labels. 
    Takes list of words, transforms to padded, uniform length vectors of word indices in embedding.
    '''
    
    data_shuffle = np.zeros([batch_size, max_sent_length])
    labels_shuffle = np.zeros([batch_size, 2])
    
    idx = np.arange(0, len(data))
    np.random.shuffle(idx)
    idx = idx[:batch_size]
    
    for i in range(batch_size):
        if type(data[idx[i]]) == str:
            sent = word_tokenize(data[idx[i]])
        elif type(data[idx[i]]) == np.str_:
            sent = word_tokenize(data[idx[i]])
        else:
            sent = []
            
        for word_idx in range(max_sent_length):
            try:
                word = str(sent[word_idx])
                data_shuffle[i][word_idx] = word2idx[word]
            except KeyError:
                data_shuffle[i][word_idx] = embed_vocab_size
            except IndexError:
                data_shuffle[i][word_idx] = embed_vocab_size
        labels_shuffle[i] = labels[idx[i]]
        
    return data_shuffle, labels_shuffle

tf.reset_default_graph()

with tf.device("/cpu:0"):
    embedding_weights = tf.Variable(tf.constant(0.0, shape=[embed_vocab_size+1, word_embed_size]),trainable=True, name="embedding_weights") 
    embedding_placeholder = tf.placeholder(tf.float32, [embed_vocab_size+1, word_embed_size])
    embedding_init = embedding_weights.assign(embedding_placeholder)

x_input = tf.placeholder(tf.int64, [None, max_sent_length])
x = tf.nn.embedding_lookup(embedding_weights, x_input)
y = tf.placeholder(tf.float32, [None, num_classes], name="y")

lstm_cell = tf.contrib.rnn.LSTMCell(lstm_units, name="lstm_cell")
lstm_cell = tf.contrib.rnn.DropoutWrapper(cell=lstm_cell, output_keep_prob=0.8)
value, _ = tf.nn.dynamic_rnn(lstm_cell, x, dtype=tf.float32)

weights = tf.Variable(tf.truncated_normal([lstm_units, num_classes]), name="weights")
bias = tf.Variable(tf.constant(0.1, shape=[num_classes]), name="bias")
value = tf.transpose(value, [1, 0, 2])
last = tf.gather(value, int(value.get_shape()[0]) - 1)
prediction = (tf.matmul(last, weights) + bias)
pred_val = tf.argmax(prediction,1)

correct = tf.equal(pred_val, tf.argmax(y,1))
accuracy = tf.reduce_mean(tf.cast(correct, tf.float32))

loss = tf.reduce_mean(tf.nn.softmax_cross_entropy_with_logits_v2(logits=prediction, labels=y))
optimizer = tf.train.AdamOptimizer().minimize(loss)

sess = tf.InteractiveSession()
sess.run(tf.global_variables_initializer())
sess.run(embedding_init, feed_dict={embedding_placeholder: embedding_matrix})

for i in range(iterations):
    batch_titles, batch_lb = nextBatch(batch_size, X_total, y_total)
    sess.run(optimizer, feed_dict = {x_input: batch_titles, y: batch_lb})
    
    if i % 500 == 0:
        print('iteration ' + str(i))
        batch_titles, batch_lb = nextBatch(500, X_train, y_train)
        print(sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb}))
        batch_titles, batch_lb = nextBatch(500, X_test, y_test)
        print(sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb}))
        batch_titles, batch_lb = nextBatch(len(X_byarticle), X_byarticle, y_byarticle)
        print(sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb}))

tf.saved_model.simple_save(sess,
            "/project/cramerus/LSTM-final",
            inputs={"Text": x_input},
            outputs={"Prediction": pred_val})

# calculate final accuracy ratings
print('Accuracy on training set:') #480,000
correct = 0.0
i = 0
while i < len(X_train):
    batch_titles, batch_lb = nextBatch(1000, np.array(X_train)[i:i+1000], y_train[i:i+1000])
    acc = sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb})
    correct += 1000 * acc
    i += 1000
print(str(correct / len(X_train)))

#print('Accuracy on holdout training set:') #120,000
#correct = 0.0
#i = 0
#while i < len(X_train_holdout):
#    batch_titles, batch_lb = nextBatch(1000, np.array(X_train_holdout)[i:i+1000], y_train_holdout[i:i+1000])
#    acc = sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb})
#    correct += 1000 * acc
#    i += 1000
#print(str(correct / len(X_train_holdout)))

print('Accuracy on test set:') #150,000
correct = 0.0
i = 0
while i < len(X_test):
    batch_titles, batch_lb = nextBatch(1000, np.array(X_test)[i:i+1000], y_test[i:i+1000])
    acc = sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb})
    correct += 1000 * acc
    i += 1000
print(str(correct / len(X_test)))

print('Accuracy on baby set:')
batch_titles, batch_lb = nextBatch(len(X_byarticle), np.array(X_byarticle), y_byarticle)
print(sess.run(accuracy, feed_dict={x_input: batch_titles, y: batch_lb}))

sess.close()

Train:
Accuracy on training set:
0.8370049995183945
Accuracy on test set:
0.5911999992529551
Accuracy on baby set:
0.5472868

Accuracy on training set:
0.5681733333071073
Accuracy on test set:
0.8237999975681305
Accuracy on baby set:
0.40465117
