==> code/07 voting-system.ipynb <==
import pickle
import numpy as np
from tensorflow.contrib import predictor
from nltk import word_tokenize
import pandas as pd
from tqdm import tqdm_notebook
from sklearn.linear_model import LogisticRegression

This is the notebook for the voting system done by the three LSTMs, each trained on a different given dataset.

All datasets are loaded, evaluated by all models, and then results are saved to another CSV file. Those examples for which all models agree correctly are saved to a new dataset to be used in a separate LSTM.

# mark votes

train = pd.read_csv('df/train_df.csv')
test = pd.read_csv('df/test_df.csv')
byarticle = pd.read_csv('df/byarticle_df.csv')

max_sent_length = 100
word_embed_size = 50
embed_vocab_size = len(word2idx.keys())
word2idx = pickle.load(open("word-embeddings/embed_total_idx.pkl", "rb" ))

def process(sent):
    # Takes text, transforms to padded, uniform length vectors of word indices in embedding.
    
    vector = np.zeros([max_sent_length])
    
    if type(sent) == str:
            sent = word_tokenize(sent)
    elif type(sent) == np.str_:
            sent = word_tokenize(sent)
    else:
        sent = []
            
    for word_idx in range(max_sent_length):
        try:
            word = str(sent[word_idx]).lower()
            vector[word_idx] = word2idx[word]
        except KeyError:
            vector[word_idx] = embed_vocab_size
        except IndexError:
            vector[word_idx] = embed_vocab_size

    return [vector]

train_pred = predictor.from_saved_model("/project/cramerus/LSTM-text-train")
test_pred = predictor.from_saved_model("/project/cramerus/LSTM-text-test")
byarticle_pred = predictor.from_saved_model("/project/cramerus/LSTM-text-byarticle")

def add_pred_cols(df):
    
    df['train_pred'] = np.nan
    df['train_pred'] = df['train_pred'].astype(object)
    df['test_pred'] = np.nan
    df['test_pred'] = df['test_pred'].astype(object)
    df['byarticle_pred'] = np.nan
    df['byarticle_pred'] = df['byarticle_pred'].astype(object)

    for idx, row in tqdm_notebook(df.iterrows(), total = df.shape[0]):
        if pd.isnull(row['text']):
            continue
        else:
            df.at[idx, 'train_pred'] = train_pred({"Sentence": process(row['text'])})['Prediction'][0]
            df.at[idx, 'test_pred'] = test_pred({"Sentence": process(row['text'])})['Prediction'][0]
            df.at[idx, 'byarticle_pred'] = byarticle_pred({"Sentence": process(row['text'])})['Prediction'][0]
            
    return df        

byarticle = add_pred_cols(byarticle)

test = add_pred_cols(test)

byarticle.to_csv('df/byarticle_df_vote.csv', index=False)

test.to_csv('df/test_df_vote.csv', index=False)

train = add_pred_cols(train)

train.to_csv('df/train_df_vote.csv', index=False)

train.head()

# explore votes

train = pd.read_csv('df/train_df_vote.csv').dropna()
test = pd.read_csv('df/test_df_vote.csv').dropna()
byarticle = pd.read_csv('df/byarticle_df_vote.csv').dropna()

train['train_pred'] = train['train_pred'].astype(int)
train['test_pred'] = train['test_pred'].astype(int)
train['byarticle_pred'] = train['byarticle_pred'].astype(int)
train['label'] = train['label'].astype(int)

test['train_pred'] = test['train_pred'].astype(int)
test['test_pred'] = test['test_pred'].astype(int)
test['byarticle_pred'] = test['byarticle_pred'].astype(int)
test['label'] = test['label'].astype(int)

byarticle['train_pred'] = byarticle['train_pred'].astype(int)
byarticle['test_pred'] = byarticle['test_pred'].astype(int)
byarticle['byarticle_pred'] = byarticle['byarticle_pred'].astype(int)
byarticle['label'] = byarticle['label'].astype(int)

combined = pd.concat([byarticle[byarticle['label']==True].sample(200),
                      byarticle[byarticle['label']==False].sample(200),
                      test[test['label']==True].sample(200),
                      test[test['label']==False].sample(200),
                      train[train['label']==True].sample(200),
                      train[train['label']==False].sample(200)])

clf = LogisticRegression()
clf.fit(combined[['train_pred', 'test_pred', 'byarticle_pred']], combined['label'])

print('byarticle score: ')
print(clf.score(byarticle[['train_pred', 'test_pred', 'byarticle_pred']], byarticle['label']))
print('train score: ')
print(clf.score(train[['train_pred', 'test_pred', 'byarticle_pred']], train['label']))
print('test score: ')
print(clf.score(test[['train_pred', 'test_pred', 'byarticle_pred']], test['label']))

comb_df = pd.concat([byarticle[byarticle['label']==True].sample(200),
                      byarticle[byarticle['label']==False].sample(200),
                      test[test['label']==True].sample(200),
                      test[test['label']==False].sample(200),
                      train[train['label']==True].sample(200),
                      train[train['label']==False].sample(200)
                    ])

clf = LogisticRegression()
clf.fit(comb_df[['train_pred', 'test_pred', 'byarticle_pred']], comb_df['label'])

print('training score: ')
print(clf.score(comb_df[['train_pred', 'test_pred', 'byarticle_pred']], comb_df['label']))
print('byarticle score: ')
print(clf.score(byarticle[['train_pred', 'test_pred', 'byarticle_pred']], byarticle['label']))
print('train score: ')
print(clf.score(train[['train_pred', 'test_pred', 'byarticle_pred']], train['label']))
print('test score: ')
print(clf.score(test[['train_pred', 'test_pred', 'byarticle_pred']], test['label']))

train_agree = train[train['label'] == train['train_pred']][train['label'] == train['test_pred']][train['label'] == train['byarticle_pred']].drop('bias', axis=1)
test_agree = test[test['label'] == test['train_pred']][test['label'] == test['test_pred']][test['label'] == test['byarticle_pred']].drop('bias', axis=1)
byarticle_agree = byarticle[byarticle['label'] == byarticle['train_pred']][byarticle['label'] == byarticle['test_pred']][byarticle['label'] == byarticle['byarticle_pred']]

count = 0
for idx, row in train_agree.iterrows():
    print(str(row['label']) + '\t' + row['text'][:100])
    count += 1
    if count > 50:
        break

train_agree.to_csv('df/train_agree_df.csv', index=False)
test_agree.to_csv('df/test_agree_df.csv', index=False)
byarticle_agree.to_csv('df/byarticle_agree_df.csv', index=False)

train_agree = train[train['label'] == train['train_pred']][train['label'] == train['test_pred']].drop('bias', axis=1)
test_agree = test[test['label'] == test['train_pred']][test['label'] == test['test_pred']].drop('bias', axis=1)

train_agree = pd.read_csv('data/agree/train_agree_df.csv')
test_agree = pd.read_csv('data/agree/test_agree_df.csv')
byarticle_agree = pd.read_csv('data/agree/byarticle_agree_df.csv')

print(len(train_agree))
print(sum(train_agree.label))

print(len(test_agree))
print(sum(test_agree.label))

print(len(byarticle_agree))
print(sum(byarticle_agree.label))

133868+28048+130

46449+13856+86

60391/162046
