==> code/08 doc2vec.ipynb <==
from gensim.utils import simple_preprocess
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.test.utils import get_tmpfile
import numpy as np
import pandas as pd
import pickle
import keras
from tqdm import tqdm_notebook
from sklearn.preprocessing import OneHotEncoder
from sklearn.linear_model import LogisticRegression

This notebook was used to train a doc2vec model using gensim, transform all texts into document vectors, and then use Keras to build a simple feedforward neural network to use on the datasets.

%cd '/data/hyperpartisan-news-detection'

# train doc2vec model

train = pd.read_csv('df/train_df.csv')['text'].dropna()
test = pd.read_csv('df/test_df.csv')['text'].dropna()
byarticle = pd.read_csv('df/byarticle_df.csv')['text'].dropna()

train = [TaggedDocument(simple_preprocess(line), [i]) for i, line in enumerate(train)]
test = [TaggedDocument(simple_preprocess(line), [i]) for i, line in enumerate(test)]
byarticle = [simple_preprocess(line) for line in byarticle]

model = Doc2Vec(vector_size=50, min_count=5, epochs=10)

model.build_vocab(train)

%time model.train(train, total_examples=model.corpus_count, epochs=model.epochs)

fname = get_tmpfile("doc2vec_model")
model.save(fname)
#model = Doc2Vec.load(fname)  # you can continue training with the loaded model!
#model.delete_temporary_training_data(keep_doctags_vectors=True, keep_inference=True)

model.save('doc2vec_model')

# convert all docs to vectors and save

fname = 'doc2vec_model'
model_load = Doc2Vec.load(fname) 

train = pd.read_csv('df/train_df.csv')
test = pd.read_csv('df/test_df.csv')
byarticle = pd.read_csv('df/byarticle_df.csv')

def add_col(df):
    df['doc2vec'] = np.nan
    df['doc2vec'] = df['doc2vec'].astype(object)
    for idx, row in tqdm_notebook(df.iterrows(), total = df.shape[0]):
        if pd.isnull(row['text']):
            continue
        else:
            preprocessed = simple_preprocess(row['text'])
        df.at[idx, 'doc2vec'] = model_load.infer_vector(preprocessed)
    return df

byarticle = add_col(byarticle)

test = add_col(test)

train = add_col(train)

def update_df(df):
    # need to have features in different columns, not as a list because pd.to_csv will save it as a string
    features = df['doc2vec'].apply(pd.Series)
    return pd.concat([df.drop('doc2vec', axis='columns'), features], axis=1)

byarticle = update_df(byarticle)
test = update_df(test)
train = update_df(train)

train.to_csv('df/train_df_d2v.csv', index=False)
test.to_csv('df/test_df_d2v.csv', index=False)
byarticle.to_csv('df/byarticle_df_d2v.csv', index=False)

# test with logistic regression

# prepare dataset
train = pd.read_csv('df/train_df_d2v.csv').dropna()
test = pd.read_csv('df/test_df_d2v.csv').dropna()
byarticle = pd.read_csv('df/byarticle_df_d2v.csv').dropna()

train[train.columns[4:]].head()

clf = LogisticRegression(verbose=1)
clf.fit(train[train.columns[4:]], train['label'])

print('byarticle score: ')
print(clf.score(byarticle[byarticle.columns[3:]], byarticle['label']))
print('train score: ')
print(clf.score(train[train.columns[4:]], train['label']))
print('test score: ')
print(clf.score(test[test.columns[4:]], test['label']))

# Keras model

from keras.models import Sequential
from keras.layers import Dense, Dropout

seed = 7
np.random.seed(seed)

# Model example 1: 1 hidden layer with 64 hidden nodes
model_d2v_01 = Sequential()
model_d2v_01.add(Dense(64, activation='relu', input_dim=50))
model_d2v_01.add(Dense(1, activation='sigmoid'))
model_d2v_01.compile(optimizer='adam',
              loss='binary_crossentropy',
              metrics=['accuracy'])

model_d2v_01.fit(train[train.columns[4:]], train['label'],
                 validation_data=(test[test.columns[4:]], test['label']),
                 epochs=5, batch_size=32, verbose=2)

model_d2v_01.evaluate(x=byarticle[byarticle.columns[3:]], y=byarticle['label'])
